% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{entropy_formula}
\alias{entropy_formula}
\title{entropy formula (used in external_validation function)}
\usage{
entropy_formula(x_vec)
}
\description{
entropy formula (used in external_validation function)
}
\keyword{internal}
