% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{function_interactive}
\alias{function_interactive}
\title{Interactive function for consecutive plots ( using dissimilarities or the silhouette widths of the observations )}
\usage{
function_interactive(evaluation_objects, max_clusters, silhouette = FALSE)
}
\description{
Interactive function for consecutive plots ( using dissimilarities or the silhouette widths of the observations )
}
\keyword{internal}
