% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{tryCatch_GMM}
\alias{tryCatch_GMM}
\title{tryCatch function to prevent armadillo errors}
\usage{
tryCatch_GMM(
  data,
  gaussian_comps,
  dist_mode,
  seed_mode,
  km_iter,
  em_iter,
  verbose,
  var_floor,
  seed
)
}
\description{
tryCatch function to prevent armadillo errors
}
\keyword{internal}
