% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{tryCatch_KMEANS_arma}
\alias{tryCatch_KMEANS_arma}
\title{tryCatch function to prevent armadillo errors in KMEANS_arma}
\usage{
tryCatch_KMEANS_arma(
  data,
  clusters,
  n_iter,
  verbose,
  seed_mode,
  CENTROIDS,
  seed
)
}
\description{
tryCatch function to prevent armadillo errors in KMEANS_arma
}
\keyword{internal}
